Move from `/details` to `/src` (on_audio_query_platform_web)
